Work around a disk controller hardware bug affecting boot.
Boot was frequently hanging on the disk controller when booted warm.
I tracked it down to a faulty design assumption in the hardware
(and the manufacturer's sample driver).  See my disk controller
hardware notes for the hardware details.  I worked around it in
DSKINIT by letting the controller's Z80 execute garbage for a
brief time before re-opening the memory window.